+++ api/src/main/java/com/iudigital/rollerspeed/RollerspeedApplication.java
@@ -1,0 +1,13 @@
+package com.iudigital.rollerspeed;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RollerspeedApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(RollerspeedApplication.class, args);
+	}
+
+}
