+++ api/src/main/java/com/iudigital/rollerspeed/config/errorhandler/GlobalControllerAdvice.java
@@ -1,0 +1,43 @@
+package com.iudigital.rollerspeed.config.errorhandler;
+
+import com.iudigital.rollerspeed.config.errorhandler.dto.CustomException;
+import com.iudigital.rollerspeed.config.errorhandler.dto.ErrorResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import java.time.LocalDateTime;
+
+@Slf4j
+@ControllerAdvice
+public class GlobalControllerAdvice {
+
+    @ExceptionHandler(CustomException.class)
+    public ResponseEntity<ErrorResponse> handleCustomException(CustomException ex) {
+        log.error("CustomException: {}", ex.getMessage(), ex);
+
+        ErrorResponse errorResponse = ErrorResponse.builder()
+                .timestamp(LocalDateTime.now())
+                .code(ex.getErrorCode().getCode())
+                .message(ex.getErrorCode().getMessage())
+                .status(ex.getStatus().name())
+                .build();
+
+        return ResponseEntity.status(ex.getStatus()).body(errorResponse);
+    }
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ErrorResponse> handleGenericException(Exception ex) {
+        log.error("Unexpected Exception: {}", ex.getMessage(), ex);
+
+        ErrorResponse errorResponse = ErrorResponse.builder()
+                .timestamp(LocalDateTime.now())
+                .code("INTERNAL_SERVER_ERROR")
+                .message("An unexpected error occurred")
+                .status(HttpStatus.INTERNAL_SERVER_ERROR.name())
+                .build();
+
+        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
+    }
+}
