+++ api/src/main/java/com/iudigital/rollerspeed/config/errorhandler/dto/ErrorResponse.java
@@ -1,0 +1,17 @@
+package com.iudigital.rollerspeed.config.errorhandler.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Builder;
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+@Builder
+public class ErrorResponse {
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS", timezone = "UTC")
+    private LocalDateTime timestamp;
+    private String code;
+    private String message;
+    private String status;
+}
