+++ api/src/main/java/com/iudigital/rollerspeed/config/jwt/provider/JwtProvider.java
@@ -1,0 +1,82 @@
+package com.iudigital.rollerspeed.config.jwt.provider;
+
+import com.iudigital.rollerspeed.config.errorhandler.dto.CustomException;
+import com.iudigital.rollerspeed.config.errorhandler.dto.ErrorCode;
+import io.jsonwebtoken.*;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+import io.jsonwebtoken.security.SignatureException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Component;
+
+import javax.crypto.SecretKey;
+import java.util.Date;
+import java.util.logging.Logger;
+
+@Component
+@RequiredArgsConstructor
+public class JwtProvider {
+
+    private static final Logger LOGGER =  Logger.getLogger(JwtProvider.class.getName());
+
+    @Value("${jwt.secret}")
+    private String secret;
+    @Value("${jwt.expiration}")
+    private Integer expiration;
+
+    public String generateToken(UserDetails userDetails) {
+        return Jwts.builder()
+                .subject(userDetails.getUsername())
+                .claim("roles", userDetails.getAuthorities())
+                .issuedAt(new Date())
+                .expiration(new Date(new Date().getTime() + expiration))
+                .signWith(getKey(secret))
+                .compact();
+    }
+
+    public Claims getClaims(String token) {
+        return Jwts.parser()
+                .verifyWith(getKey(secret))
+                .build()
+                .parseSignedClaims(token)
+                .getPayload();
+    }
+
+    public String getSubject(String token) {
+        return Jwts.parser()
+                .verifyWith(getKey(secret))
+                .build()
+                .parseSignedClaims(token)
+                .getPayload()
+                .getSubject();
+    }
+
+    public boolean validate(String token){
+        try {
+            Jwts.parser()
+                    .verifyWith(getKey(secret))
+                    .build()
+                    .parseSignedClaims(token)
+                    .getPayload()
+                    .getSubject();
+            return true;
+        } catch (ExpiredJwtException e) {
+            throw new CustomException(ErrorCode.TOKEN_EXPIRED);
+        } catch (UnsupportedJwtException e) {
+            throw new CustomException(ErrorCode.TOKEN_UNSUPPORTED);
+        } catch (MalformedJwtException e) {
+            throw new CustomException(ErrorCode.TOKEN_MALFORMED);
+        } catch (SignatureException e) {
+            throw new CustomException(ErrorCode.TOKEN_BAD_SIGNATURE);
+        } catch (IllegalArgumentException e) {
+            throw new CustomException(ErrorCode.TOKEN_ILLEGAL_ARGUMENT);
+        }
+    }
+
+    private SecretKey getKey(String secret) {
+        byte[] secretBytes = Decoders.BASE64URL.decode(secret);
+        return Keys.hmacShaKeyFor(secretBytes);
+    }
+}
