+++ api/src/main/java/com/iudigital/rollerspeed/controller/user/login/LoginRequest.java
@@ -1,0 +1,5 @@
+package com.iudigital.rollerspeed.controller.user.login;
+
+public record LoginRequest(String userName,
+                           String password
+) {}
