+++ api/src/main/java/com/iudigital/rollerspeed/model/user/login/Role.java
@@ -1,0 +1,5 @@
+package com.iudigital.rollerspeed.model.user.login;
+
+public enum Role {
+    USER, ADMIN;
+}
