+++ api/src/main/java/com/iudigital/rollerspeed/service/UserService.java
@@ -1,0 +1,7 @@
+package com.iudigital.rollerspeed.service;
+
+import com.iudigital.rollerspeed.model.user.login.UserLogin;
+
+public interface UserService {
+    UserLogin login(String username, String password);
+}
