--- api/src/main/java/com/rollerspeed/rollerspeed/RollerspeedApplication.java
@@ -1,13 +1,0 @@
-package com.rollerspeed.rollerspeed;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RollerspeedApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RollerspeedApplication.class, args);
-	}
-
-}
