--- api/src/main/java/com/rollerspeed/rollerspeed/config/errorhandler/DTO/ErrorResponse.java
@@ -1,17 +1,0 @@
-package com.rollerspeed.rollerspeed.config.errorhandler.DTO;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.Builder;
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-@Builder
-public class ErrorResponse {
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS", timezone = "UTC")
-    private LocalDateTime timestamp;
-    private String code;
-    private String message;
-    private String status;
-}
