--- api/src/main/java/com/rollerspeed/rollerspeed/model/user/login/Role.java
@@ -1,5 +1,0 @@
-package com.rollerspeed.rollerspeed.model.user.login;
-
-public enum Role {
-    USER, ADMIN;
-}
