--- api/src/main/java/com/rollerspeed/rollerspeed/service/UserService.java
@@ -1,9 +1,0 @@
-package com.rollerspeed.rollerspeed.service;
-
-import com.rollerspeed.rollerspeed.model.user.login.UserLogin;
-
-import java.util.Optional;
-
-public interface UserService {
-    UserLogin login(String username, String password);
-}
