+++ api/src/test/java/com/iudigital/rollerspeed/RollerspeedApplicationTests.java
@@ -1,0 +1,13 @@
+package com.iudigital.rollerspeed;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class RollerspeedApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
