--- api/src/test/java/com/rollerspeed/rollerspeed/RollerspeedApplicationTests.java
@@ -1,13 +1,0 @@
-package com.rollerspeed.rollerspeed;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RollerspeedApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
